+++ guillotina/contrib/dbusers/install.py
@@ -4,6 +4,8 @@ from guillotina.content import create_content_in_container
 from guillotina.interfaces import ILayers
 from guillotina.utils import get_authenticated_user_id
 from guillotina.utils import get_registry
+from guillotina.event import notify
+from guillotina.events import ObjectAddedEvent
 
 
 USERS_LAYER = "guillotina.contrib.dbusers.interfaces.IDBUsersLayer"
@@ -16,12 +18,14 @@ class DBUsersAddon(Addon):
         registry = await get_registry()
         registry.for_interface(ILayers)["active_layers"] |= {USERS_LAYER}
         user = get_authenticated_user_id()
-        await create_content_in_container(
+        user_manager = await create_content_in_container(
             site, "UserManager", "users", creators=(user,), title="Users", check_constraints=False
         )
-        await create_content_in_container(
+        await notify(ObjectAddedEvent(user_manager))
+        group_manager = await create_content_in_container(
             site, "GroupManager", "groups", creators=(user,), title="Groups", check_constraints=False
         )
+        await notify(ObjectAddedEvent(group_manager))
 
     @classmethod
     async def uninstall(self, site, request):